requirements.txt
matplotlib
kornia
torch
torchvision
facenet-pytorch
pandas
scikit-image
googledrivedownloader
tqdm
Pillow
